ansible/roles.py: fix detecting old commit against remote
Core issue was use of `rev-parse @{u}` which represents default upstream commit.
Blecause we use `http-origin` to speed up the fetch process when we call
`rev-parse @{u}` to check upstream state we check `origin` upstream commit.
This is wrong because we call `fetch` only on `http-origin` so using @{u} is wrong.

Signed-off-by: Jakub Sokołowski <jakub@status.im>